change reindex day (Sat,Sun) better handle errors

+++ openprocurement/search/index/__init__.py
@@ -69,20 +69,23 @@ class BaseIndex:
         return name
 
     def delete_index(self, name):
-        # TODO: real delete index
-        return
+        index_key = self.__index_name__
+        index_key_prev = "{}.prev".format(index_key)
+        self.engine.set_index(index_key_prev, name)
 
     def set_current(self, name):
-        logger.warning("Set current, index %s", name)
+        index_key = self.__index_name__
         old_index = self.current_index
-        index_key, suffix = name.rsplit('_', 1)
-        self.engine.set_index(index_key, name)
-        assert(self.current_index == name)
+        if name != old_index:
+            logger.info("Change current index %s -> %s",
+                old_index, name)
+            self.engine.set_index(index_key, name)
+            assert(self.current_index == name)
+            self.delete_index(old_index)
         # remove index.next key
         index_key_next = "{}.next".format(index_key)
         if self.engine.get_index(index_key_next) == name:
             self.engine.set_index(index_key_next, '')
-        self.delete_index(old_index)
 
     def test_exists(self, index_name, info):
         return self.engine.test_exists(index_name, info)
@@ -102,15 +105,14 @@ class BaseIndex:
         return self.engine.index_item(index_name, item)
 
     def index_source(self, index_name=None, reset=False):
-        source = self.source
         if reset:
-            source.reset()
+            self.source.reset()
 
         if not index_name:
             index_name = self.current_index
 
         if not index_name and self.engine.slave_mode:
-            self.engine.heartbeat(source)
+            self.engine.heartbeat(self.source)
             index_name = self.current_index
 
         if not index_name:
@@ -119,21 +121,21 @@ class BaseIndex:
 
         index_count = 0
         total_count = 0
-        while self.engine.heartbeat(source):
-            items_list = source.items()
+        while self.engine.heartbeat(self.source):
+            items_list = self.source.items()
             iter_count = 0
             for info in items_list:
                 if not self.test_exists(index_name, info):
-                    item = source.get(info)
+                    item = self.source.get(info)
                     if self.index_item(index_name, item):
                         index_count += 1
                 iter_count += 1
                 total_count += 1
-                # update heartbeat for long indexing
+                # update heartbeat for long indexing (ocds)
                 if iter_count % 1000 == 0:
                     logger.info("[%s] Fetched %d indexed %d",
                         index_name, total_count, index_count)
-                    self.engine.heartbeat(source)
+                    self.engine.heartbeat(self.source)
                     sleep(1)
             # break on empty set
             if not iter_count:
@@ -149,7 +151,7 @@ class BaseIndex:
     def process(self, allow_reindex = True):
         if self.need_reindex() and allow_reindex:
             index_name = self.new_index()
-            logger.warning("Starting full re-index, new index %s",
+            logger.info("Starting full re-index, new index %s",
                 index_name)
             self.index_source(index_name, reset=True)
             self.finish_index(index_name)
